requirements.txt
antlr4-python3-runtime~=4.9.1
Pillow~=7.0.0
python-telegram-bot~=13.1